// src/types/resources/Encounter.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TEncounterStatusHistory } from '../partials/EncounterStatusHistory';
import { TCoding } from '../partials/Coding';
import { TEncounterClassHistory } from '../partials/EncounterClassHistory';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TEncounterParticipant } from '../partials/EncounterParticipant';
import { TPeriod } from '../partials/Period';
import { TQuantity } from '../partials/Quantity';
import { TEncounterDiagnosis } from '../partials/EncounterDiagnosis';
import { TEncounterHospitalization } from '../partials/EncounterHospitalization';
import { TEncounterLocation } from '../partials/EncounterLocation';

export type TEncounter = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  status: string;
  statusHistory?: TEncounterStatusHistory[];
  class_: TCoding;
  classHistory?: TEncounterClassHistory[];
  type?: TCodeableConcept[];
  serviceType?: TCodeableConcept;
  priority?: TCodeableConcept;
  subject?: TReference;
  episodeOfCare?: TReference[];
  basedOn?: TReference[];
  participant?: TEncounterParticipant[];
  appointment?: TReference[];
  period?: TPeriod;
  length?: TQuantity;
  reasonCode?: TCodeableConcept[];
  reasonReference?: TReference[];
  diagnosis?: TEncounterDiagnosis[];
  account?: TReference[];
  hospitalization?: TEncounterHospitalization;
  location?: TEncounterLocation[];
  serviceProvider?: TReference;
  partOf?: TReference;
};